Объединил две конфигурации s3 в одну.
